Fixed comments.

+++ src/function/s21_ceil.c
@@ -18,7 +18,7 @@ long double s21_ceil(double x) {
   else {
     mod = (double)s21_fmod(x, 1.0);
     result = x - mod;
-    if ((s21_fabs((double)mod) > 0) && x > 0) result++;
+    if ((s21_fabs(mod) > 0) && x > 0) result++;
   }
   return result;
 }
